final pass for TOML parser + importer
- refactoring pending
- only few unit tests added yet

Whenever TOML file changes, upon start program will update the database entry to match with TOML file config.
This makes use of an identifier field which should never change for a particular workflow, as that's used to identify whether this would be inserted as a new workflow or update an existing workflow. It should also be unique for each workflow that's defined.
In order to identify workflow step changes, which are more complicated, we maintain a hash, which when changed, triggers the deletion of old data and fresh import of new data.
Defines some extra get/insert/update functions which will probably be used during refactoring, even though some of them are not used at point of this commit.